move <Photo> to its own file
